Simpler FHSS Generation (#830)
* Simplified FHSS sequence generation

* Simpler FHSS generation

* define regulatory domain

* Fixed random number generator. Changed FHSS sequence to be a multiple of the number of entries in size.

* Updated variable names for consistency

* Updated debugging

* Fixed test, added comments, fixed broken references

+++ src/lib/FHSS/random.h
@@ -5,7 +5,7 @@
 // the max value returned by rng
 #define RNG_MAX 0x7FFF
 
-uint32_t rng(void);
+uint16_t rng(void);
 
 void rngSeed(uint32_t newSeed);
 // 0..255 returned
@@ -13,5 +13,5 @@ uint8_t rng8Bit(void);
 // 0..31 returned
 uint8_t rng5Bit(void);
 
-// returns 0 <= x < n where n <= 256
-uint32_t rngN(uint32_t upper);
+// returns 0 <= x < upper where n < 256
+uint8_t rngN(uint8_t upper);
